dom/devicestorage/test/test_watchOther.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html> <!--
https://bugzilla.mozilla.org/show_bug.cgi?id=717103
-->
<head>
  <title>Test for the device storage API </title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="devicestorage_common.js"></script>

<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=717103">Mozilla Bug 717103</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" type="text/javascript">

devicestorage_setup();

var gFileName = randomFilename(20) + ".png"

function addSuccess(e) {
}

function addError(e) {
  ok(false, "addError was called : " + e.target.error.name);
  devicestorage_cleanup();
}

function onChange(e) {

  dump("we saw: " + e.path + " " + e.reason + "\n");

  var filename = e.path;
  if (filename[0] == "/") {
    // We got /storageName/prefix/filename
    // Remove the storageName (this shows up on FirefoxOS)
    filename = filename.substring(1); // Remove leading slash
    var slashIndex = filename.indexOf("/");
    if (slashIndex >= 0) {
      filename = filename.substring(slashIndex + 1); // Remove storageName
    }
  }
  if (filename == gFileName) {
    ok(true, "we saw the file get created");
    storage.removeEventListener("change", onChange);
    devicestorage_cleanup();
  }
  else {
    // we may see other file changes during the test, and
    // that is completely ok
  }
}

function onChangeFail(e) {
  dump("onChangeFail: " + e.path + " " + e.reason + "\n");
  ok(false, "We should never see any changes");
}

var storage = navigator.getDeviceStorage("pictures");
ok(storage, "Should have storage");
storage.addEventListener("change", onChange);

var storageOther = navigator.getDeviceStorage("music");
ok(storageOther, "Should have storage");
storageOther.addEventListener("change", onChangeFail);

request = storage.addNamed(createRandomBlob('image/png'), gFileName);
ok(request, "Should have a non-null request");

request.onsuccess = addSuccess;
request.onerror = addError;

</script>
</pre>
</body>
</html>
